requirements.txt
allure-pytest>=2.9.43
pytest>=6.2.4
